Added fix for focus behaviour in pages with a .form-section-preview
The current behaviour for the focus state when providing an update to a report is going directly into the first input after the .form-section-preview and its content, making it difficult for users using assistive devices to edit an "update" once they have pressed the "continue" button, because they would need to travel across the whole page in order to reach "Edit your update" button. With this fix the focus state goes directly to the content inside ".form-section-preview" if this element exist.